backend/app/requirements.txt
requests==2.32.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9
tenacity==8.2.3
openai==1.14.0
langchain==0.1.4
PyMuPDF==1.23.14
types-requests==2.31.0.10
mypy==1.7.1
pytest==7.4.3
black==23.12.0
isort==5.13.0
flake8==6.1.0
python-logging-loki==0.3.1